// Warehousing.UI/src/app/admin/constants/enums/permissions.enum.ts
export enum PermissionsEnum {

    //High level permissions
    VIEW_ADMIN = 'VIEW_ADMIN',

    // 🔹 Products Module
    VIEW_PRODUCTS = 'VIEW_PRODUCTS',
    ADD_PRODUCT = 'ADD_PRODUCT',
    EDIT_PRODUCT = 'EDIT_PRODUCT',
    DELETE_PRODUCT = 'DELETE_PRODUCT',
    PRINT_PRODUCTS = 'PRINT_PRODUCTS',

    // 🔹 Users Module
    VIEW_USERS = 'VIEW_USERS',
    ADD_USER = 'ADD_USER',
    EDIT_USER = 'EDIT_USER',
    DELETE_USER = 'DELETE_USER',
    RESET_USER_PASSWORD = 'RESET_USER_PASSWORD',
    ASSIGN_ROLES_TO_USER = 'ASSIGN_ROLES_TO_USER',

    // 🔹 Roles Module
    VIEW_ROLES = 'VIEW_ROLES',
    ADD_ROLE = 'ADD_ROLE',
    EDIT_ROLE = 'EDIT_ROLE',
    DELETE_ROLE = 'DELETE_ROLE',
    ASSIGN_PERMISSIONS_TO_ROLE = 'ASSIGN_PERMISSIONS_TO_ROLE',

    // 🔹 Customers Module
    VIEW_CUSTOMERS = 'VIEW_CUSTOMERS',
    ADD_CUSTOMER = 'ADD_CUSTOMER',
    EDIT_CUSTOMER = 'EDIT_CUSTOMER',
    DELETE_CUSTOMER = 'DELETE_CUSTOMER',
    PRINT_CUSTOMERS = 'PRINT_CUSTOMERS',

    // 🔹 Suppliers Module
    VIEW_SUPPLIERS = 'VIEW_SUPPLIERS',
    ADD_SUPPLIER = 'ADD_SUPPLIER',
    EDIT_SUPPLIER = 'EDIT_SUPPLIER',
    DELETE_SUPPLIER = 'DELETE_SUPPLIER',
    PRINT_SUPPLIERS = 'PRINT_SUPPLIERS',

    // 🔹 Purchase Orders Module
    VIEW_PURCHASE_ORDERS = 'VIEW_PURCHASE_ORDERS',
    ADD_PURCHASE_ORDER = 'ADD_PURCHASE_ORDER',
    EDIT_PURCHASE_ORDER = 'EDIT_PURCHASE_ORDER',
    DELETE_PURCHASE_ORDER = 'DELETE_PURCHASE_ORDER',
    COMPLETE_PURCHASE_ORDER = 'COMPLETE_PURCHASE_ORDER',
    PRINT_PURCHASE_ORDER = 'PRINT_PURCHASE_ORDER',

    // 🔹 Sale Orders Module
    VIEW_SALE_ORDERS = 'VIEW_SALE_ORDERS',
    ADD_SALE_ORDER = 'ADD_SALE_ORDER',
    EDIT_SALE_ORDER = 'EDIT_SALE_ORDER',
    DELETE_SALE_ORDER = 'DELETE_SALE_ORDER',
    COMPLETE_SALE_ORDER = 'COMPLETE_SALE_ORDER',
    PRINT_SALE_ORDER = 'PRINT_SALE_ORDER',

    // 🔹 Inventory Report Module
    VIEW_INVENTORY_REPORT = 'VIEW_INVENTORY_REPORT',
    PRINT_INVENTORY_REPORT = 'PRINT_INVENTORY_REPORT',

    //To view Company settings
    VIEW_SETTINGS = 'VIEW_SETTINGS',
    EDIT_SETTINGS = 'EDIT_SETTINGS',

    // 🔹 New: Category Module
    VIEW_CATEGORIES = 'VIEW_CATEGORIES',
    ADD_CATEGORY = 'ADD_CATEGORY',
    EDIT_CATEGORY = 'EDIT_CATEGORY',
    DELETE_CATEGORY = 'DELETE_CATEGORY',
    PRINT_CATEGORIES = 'PRINT_CATEGORIES',

    // 🔹 New: Unit Module
    VIEW_UNITS = 'VIEW_UNITS',
    ADD_UNIT = 'ADD_UNIT',
    EDIT_UNIT = 'EDIT_UNIT',
    DELETE_UNIT = 'DELETE_UNIT',
    PRINT_UNITS = 'PRINT_UNITS',

    // 🔹 New: Store Module
    VIEW_STORES = 'VIEW_STORES',
    ADD_STORE = 'ADD_STORE',
    EDIT_STORE = 'EDIT_STORE',
    DELETE_STORE = 'DELETE_STORE',
    PRINT_STORES = 'PRINT_STORES'
}